+++ include/ncglm.h
@@ -37,7 +37,7 @@ extern "C" {
     int glm_read_event_arrays(int ncid, size_t *nevent, int *event_id,
                               float *time_offset, float *lat, float *lon,
                               float *energy, int *parent_group_id);
-    int glm_read_group_arrays(int ncid, size_t *ngroups, unsigned int *time_offset,
+    int glm_read_group_arrays(int ncid, size_t *ngroups, float *time_offset,
                               float *lat, float *lon, float *energy, float *area,
                               unsigned int *parent_flash_id, short *quality_flag);
     int glm_read_flash_arrays(int ncid, size_t *nflash,
